requirements.txt
# Core Framework (FastAPI, SQLAlchemy, Alembic)
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.13.0

# Database (SQLite for development)
# psycopg2-binary==2.9.9  # Commented out for Windows compatibility
# geoalchemy2==0.14.2     # Commented out for Windows compatibility

# Machine Learning (simplified for Windows)
scikit-learn==1.3.2
tensorflow==2.15.0
pandas==2.1.4
numpy==1.24.3
scipy==1.11.4
xgboost==2.0.2
lightgbm==4.1.0

# Data Processing
requests==2.31.0
beautifulsoup4==4.12.2
schedule==1.2.0
celery==5.3.4
redis==5.0.1

# Geospatial (simplified for Windows)
folium==0.15.1
# geopandas==0.14.1  # Commented out for Windows compatibility
shapely==2.0.2
# rasterio==1.3.9    # Commented out for Windows compatibility

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# API and Web
httpx==0.25.2
jinja2==3.1.2
python-multipart==0.0.6

# Monitoring and Logging
prometheus-client==0.19.0
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Environment
python-dotenv==1.0.0
pyyaml==6.0.1

# Notifications
twilio==8.11.0
sendgrid==6.11.0

# Time series
statsmodels==0.14.0
prophet==1.1.5
